Created EditboxLines
